Refactor lorawan concentrator service
