+++ src/AzureIoTHub.Portal.Tests.Unit/Server/Controllers/v1.0/LoRaWAN/LoRaWANConcentratorsControllerTest.cs
@@ -16,7 +16,6 @@ namespace AzureIoTHub.Portal.Tests.Unit.Server.Controllers.v1._0.LoRaWAN
     using FluentAssertions;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.AspNetCore.Mvc.Routing;
-    using Microsoft.Azure.Devices;
     using Microsoft.Azure.Devices.Common.Exceptions;
     using Microsoft.Azure.Devices.Shared;
     using Microsoft.Extensions.Logging;
@@ -336,25 +335,8 @@ namespace AzureIoTHub.Portal.Tests.Unit.Server.Controllers.v1._0.LoRaWAN
                 RouterConfig = new RouterConfig()
             };
 
-            var twin = new Twin
-            {
-                DeviceId = concentrator.DeviceId,
-            };
-
-            var device = new Device(concentrator.DeviceId);
-
-            _ = this.mockRouterConfigManager.Setup(x => x.GetRouterConfig(It.Is<string>(c => c == concentrator.LoraRegion)))
-                .ReturnsAsync(concentrator.RouterConfig);
-            _ = this.mockConcentratorTwinMapper.Setup(x => x.UpdateTwin(It.Is<Twin>(c => c.DeviceId == twin.DeviceId), It.Is<Concentrator>(c => c.DeviceId == concentrator.DeviceId)));
-
-            _ = this.mockDeviceService.Setup(x => x.GetDevice(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ReturnsAsync(device);
-            _ = this.mockDeviceService.Setup(x => x.UpdateDevice(It.Is<Device>(c => c.Id == concentrator.DeviceId)))
-                .ReturnsAsync(device);
-            _ = this.mockDeviceService.Setup(x => x.GetDeviceTwin(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ReturnsAsync(twin);
-            _ = this.mockDeviceService.Setup(x => x.UpdateDeviceTwin(It.Is<Twin>(c => c.DeviceId == concentrator.DeviceId)))
-                .ReturnsAsync(twin);
+            _ = this.mockLoRaWANConcentratorService.Setup(x => x.UpdateDeviceAsync(It.Is<Concentrator>(c => c == concentrator)))
+                .ReturnsAsync(true);
 
             // Act
             var result = await concentratorController.UpdateDeviceAsync(concentrator);
@@ -366,104 +348,6 @@ namespace AzureIoTHub.Portal.Tests.Unit.Server.Controllers.v1._0.LoRaWAN
             Assert.IsNotNull(okObjectResult);
             Assert.AreEqual(200, okObjectResult.StatusCode);
         }
-
-        [Test]
-        public async Task WhenGetDeviceThrowAnErrorUpdateDeviceAsyncWithValidArgumentShouldThrowInernalServalErrorException()
-        {
-            // Arrange
-            var concentratorController = CreateLoRaWANConcentratorsController();
-            var concentrator = new Concentrator
-            {
-                DeviceId = "4512457896451156",
-                LoraRegion = Guid.NewGuid().ToString(),
-                IsEnabled = true,
-                RouterConfig = new RouterConfig()
-            };
-
-            _ = this.mockDeviceService.Setup(x => x.GetDevice(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ThrowsAsync(new InternalServerErrorException(""));
-
-            // Act
-            var result = async () => await concentratorController.UpdateDeviceAsync(concentrator);
-
-            // Assert
-            _ = await result.Should().ThrowAsync<InternalServerErrorException>();
-
-            this.mockRepository.VerifyAll();
-        }
-
-        [Test]
-        public async Task WhenUpdateDeviceThrowAnErrorUpdateDeviceAsyncWithValidArgumentShouldThrowInernalServalErrorException()
-        {
-            // Arrange
-            var concentratorController = CreateLoRaWANConcentratorsController();
-            var concentrator = new Concentrator
-            {
-                DeviceId = "4512457896451156",
-                LoraRegion = Guid.NewGuid().ToString(),
-                IsEnabled = true,
-                RouterConfig = new RouterConfig()
-            };
-
-            var device = new Device(concentrator.DeviceId);
-
-            _ = this.mockDeviceService.Setup(x => x.GetDevice(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ReturnsAsync(device);
-
-            _ = this.mockDeviceService.Setup(x => x.UpdateDevice(It.Is<Device>(c => c.Id == concentrator.DeviceId)))
-                .ThrowsAsync(new InternalServerErrorException(""));
-
-            // Act
-            var result = async () => await concentratorController.UpdateDeviceAsync(concentrator);
-
-            // Assert
-            _ = await result.Should().ThrowAsync<InternalServerErrorException>();
-
-            this.mockRepository.VerifyAll();
-        }
-
-        [Test]
-        public async Task WhenUpdateDeviceTwinThrowAnErrorUpdateDeviceAsyncWithValidArgumentShouldThrowInernalServalErrorException()
-        {
-            // Arrange
-            var concentratorController = CreateLoRaWANConcentratorsController();
-            var concentrator = new Concentrator
-            {
-                DeviceId = "4512457896451156",
-                LoraRegion = Guid.NewGuid().ToString(),
-                IsEnabled = true,
-                RouterConfig = new RouterConfig()
-            };
-
-            var twin = new Twin
-            {
-                DeviceId = concentrator.DeviceId,
-            };
-
-            var device = new Device(concentrator.DeviceId);
-
-            _ = this.mockRouterConfigManager.Setup(x => x.GetRouterConfig(It.Is<string>(c => c == concentrator.LoraRegion)))
-                .ReturnsAsync(concentrator.RouterConfig);
-            _ = this.mockConcentratorTwinMapper.Setup(x => x.UpdateTwin(It.Is<Twin>(c => c.DeviceId == twin.DeviceId), It.Is<Concentrator>(c => c.DeviceId == concentrator.DeviceId)));
-
-            _ = this.mockDeviceService.Setup(x => x.GetDevice(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ReturnsAsync(device);
-            _ = this.mockDeviceService.Setup(x => x.UpdateDevice(It.Is<Device>(c => c.Id == concentrator.DeviceId)))
-                .ReturnsAsync(device);
-            _ = this.mockDeviceService.Setup(x => x.GetDeviceTwin(It.Is<string>(c => c == concentrator.DeviceId)))
-                .ReturnsAsync(twin);
-            _ = this.mockDeviceService.Setup(x => x.UpdateDeviceTwin(It.Is<Twin>(c => c.DeviceId == concentrator.DeviceId)))
-                .ThrowsAsync(new InternalServerErrorException(""));
-
-            // Act
-            var result = async () => await concentratorController.UpdateDeviceAsync(concentrator);
-
-            // Assert
-            _ = await result.Should().ThrowAsync<InternalServerErrorException>();
-
-            this.mockRepository.VerifyAll();
-        }
-
         [Test]
         public async Task DeleteDeviceAsync()
         {
